--- jobs/sweep_schedulers/Dockerfile
@@ -1,9 +1,0 @@
-from python:3.10-bullseye
-
-RUN pip install --upgrade pip
-RUN pip install wandb 
-
-COPY . .
-
-# Fill me in with "wandb_scheduler.py" or "optuna_scheduler.py" etc.
-ENTRYPOINT [ "python", <PYTHON FILE> ]
