+++ jobs/sweep_schedulers/wandb_scheduler/example_configs/basic.yaml
@@ -1,0 +1,18 @@
+# basic.yaml
+scheduler:
+   job: 'wandb/jobs/Wandb Scheduler Image Job:latest'
+   resource: local-container  # required for image jobs
+
+   settings:
+      method: bayes  # required to specify method here for Wandb scheduler 
+
+job: 'wandb/jobs/Hello World 2:latest'
+run_cap: 5
+metric:
+  goal: maximize
+  name: val_acc
+
+parameters:
+   param1:
+      min: 0
+      max: 10
